refactor: move comment reads to comment/api/comment
